Added variable sterivex number in system

+++ include/Main_functions.h
@@ -6,9 +6,9 @@
 void step_dive(int _depth);
 void step_purge();
 void step_fill_container();
-void step_sampling(int num_sterivex);
+void step_sampling(uint8_t num_sterivex);
 void step_rewind();
-void step_dry(int num_sterivex);
+void step_dry(uint8_t num_sterivex);
 
 
 #endif
